update github v18 action to build together both project
