+++ DoseRateEditor.sln
@@ -1,25 +1,47 @@
-﻿Microsoft Visual Studio Solution File, Format Version 12.00
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.13.35828.75 d17.13
+VisualStudioVersion = 17.13.35828.75
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DoseRateEditor", "DoseRateEditor\DoseRateEditor.csproj", "{0C6686F0-5766-4D2A-85FB-88864066B7CD}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AOS_VirtualCones_MCB", "VirtualCones\AOS_VirtualCones_MCB.csproj", "{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "GithubActions", "GithubActions", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
+	ProjectSection(SolutionItems) = preProject
+		.github\workflows\pipeline_15.6.yaml = .github\workflows\pipeline_15.6.yaml
+		.github\workflows\pipeline_16.1.yaml = .github\workflows\pipeline_16.1.yaml
+		.github\workflows\pipeline_17.0.yaml = .github\workflows\pipeline_17.0.yaml
+		.github\workflows\pipeline_18.0.yaml = .github\workflows\pipeline_18.0.yaml
+		.github\workflows\static.yml = .github\workflows\static.yml
+		.github\workflows\Update-AssemblyInfo.ps1 = .github\workflows\Update-AssemblyInfo.ps1
+		.github\workflows\Update-EsapiHintPaths.ps1 = .github\workflows\Update-EsapiHintPaths.ps1
+	EndProjectSection
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
 		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{0C6686F0-5766-4D2A-85FB-88864066B7CD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{0C6686F0-5766-4D2A-85FB-88864066B7CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0C6686F0-5766-4D2A-85FB-88864066B7CD}.Debug|x64.ActiveCfg = Debug|x64
+		{0C6686F0-5766-4D2A-85FB-88864066B7CD}.Debug|x64.Build.0 = Debug|x64
 		{0C6686F0-5766-4D2A-85FB-88864066B7CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0C6686F0-5766-4D2A-85FB-88864066B7CD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0C6686F0-5766-4D2A-85FB-88864066B7CD}.Release|x64.ActiveCfg = Release|x64
+		{0C6686F0-5766-4D2A-85FB-88864066B7CD}.Release|x64.Build.0 = Release|x64
+		{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}.Debug|Any CPU.Build.0 = Debug|x64
+		{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}.Debug|x64.ActiveCfg = Debug|x64
+		{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}.Debug|x64.Build.0 = Debug|x64
+		{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}.Release|Any CPU.ActiveCfg = Release|x64
+		{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}.Release|Any CPU.Build.0 = Release|x64
+		{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}.Release|x64.ActiveCfg = Release|x64
+		{9CD21C0B-6792-46D5-BFA2-E1BB322BE15E}.Release|x64.Build.0 = Release|x64
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
