+++ VirtualCones/Properties/AssemblyInfo.cs
@@ -17,7 +17,7 @@ using System.Windows;
 [assembly: AssemblyCulture("")]
 
 // Expiration date for the assembly
-[assembly: AssemblyExpirationDate("2025-12-31")]
+[assembly: AssemblyExpirationDate("1/1/2026")]
 
 // Setting ComVisible to false makes the types in this assembly not visible
 // to COM components.  If you need to access a type in this assembly from
