+++ VirtualCones/packages.config
@@ -5,9 +5,9 @@
   <package id="Microsoft.Bcl.AsyncInterfaces" version="9.0.1" targetFramework="net472" />
   <package id="Microsoft.NETFramework.ReferenceAssemblies" version="1.0.3" targetFramework="net472" developmentDependency="true" />
   <package id="Microsoft.NETFramework.ReferenceAssemblies.net472" version="1.0.3" targetFramework="net472" developmentDependency="true" />
-  <package id="OxyPlot.Core" version="2.2.0" targetFramework="net472" />
-  <package id="OxyPlot.Wpf" version="2.2.0" targetFramework="net472" />
-  <package id="OxyPlot.Wpf.Shared" version="2.2.0" targetFramework="net472" />
+  <package id="OxyPlot.Core" version="2.1.0" targetFramework="net48" />
+  <package id="OxyPlot.Wpf" version="2.1.0" targetFramework="net48" />
+  <package id="OxyPlot.Wpf.Shared" version="2.1.0" targetFramework="net48" />
   <package id="Prism.Core" version="8.0.0.1909" targetFramework="net472" />
   <package id="System.Buffers" version="4.6.0" targetFramework="net472" />
   <package id="System.Collections.Immutable" version="9.0.1" targetFramework="net472" />
